requirements.txt
Flask
PyPDF2
werkzeug
google-generativeai
python-dotenv
